requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h7a1dbc1_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py37_0
freetype=2.9.1=ha9979f8_1
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
intel-openmp=2020.1=216
jpeg=9b=hb83a4c4_2
kiwisolver=1.2.0=py37h74a9793_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h64f37c6_0
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.1=py37h47e9c7a_0
ninja=1.9.0=py37h74a9793_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1g=he774522_0
pandas=1.0.3=py37h47e9c7a_0
pillow=7.1.2=py37hcc1f983_0
pip=20.0.2=py37_3
py-opencv=3.4.2=py37hc319ecb_0
pycparser=2.20=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
pytorch=1.2.0=py3.7_cuda100_cudnn7_1
pytz=2020.1=py_0
pyyaml=5.3.1=pypi_0
qt=5.9.7=vc14h73c81de_0
setuptools=46.4.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.14.0=py37_0
sqlite=3.31.1=h2a8f88b_1
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.0=py37_cu100
tornado=6.0.4=py37he774522_1
vc=14.1=h21ff451_1
vs2015_runtime=14.16.27012=hf0eaf9b_2
vs2017_runtime=15.4.27004.2010=1
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
yacs=0.1.7=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.3.7=h508b16e_0
